requirements.txt
sentence-transformers==4.1.0
langchain==0.3.25
langchain-community==0.3.24
streamlit==1.45.1
pypdf2==3.0.1
faiss-cpu==1.11.0
InstructorEmbedding==1.0.1
python-dotenv==1.1.0
openai==1.79.0